チェック追加

+++ phalp/utils/io.py
@@ -44,6 +44,9 @@ class IO_Manager():
 
 
         if(source_path.endswith(".mp4")):
+            if not os.path.exists(source_path):
+                raise Exception(f"Video file does not exist: {source_path}")
+
             # find a proper video name based on the source path
             video_name = source_path.split('/')[-1].split('.')[0]
             os.system("rm -rf " + self.cfg.video.output_dir + "/_DEMO/" + video_name + "/img/")
